controllers views models style: implement warning banner
The warning implementation only uses a string in which every alert
is concatened. It starts empty (and so doesn't show on page) and
appends step by step with the usage of add_warning. reset_warning
has also been implemented to clean the warning buffer after
displaying it so it won't show twice.